style(design-system): init adjust component architecture
