+++ app/src/main/java/com/yogiveloper/yonewsai/MainActivity.kt
@@ -9,13 +9,14 @@ import dagger.hilt.android.AndroidEntryPoint
 import androidx.compose.material3.Surface
 import androidx.navigation.compose.rememberNavController
 import com.yogiveloper.yonewsai.core.NavGraph
+import com.yogiveloper.yonewsai.ui.theme.YoNewsAiTheme
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            MaterialTheme { // NewsAppTheme
+            YoNewsAiTheme { // NewsAppTheme // MaterialTheme
                 Surface(color = MaterialTheme.colorScheme.background) {
                     val navController = rememberNavController()
                     NavGraph(navController)
