+++ app/src/main/java/com/yogiveloper/yonewsai/ui/molecules/EmptyState.kt
@@ -1,0 +1,2 @@
+package com.yogiveloper.yonewsai.ui.molecules
+
